requirements.txt
openai>=1.30.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
